This eliminates a race condition on hid_close(). Previous to this patch, hid_close() would hang on pthread_join() if hid_close() was called while the reading thread was executing read_callback() _before_ the call to libusb_submit_transfer(). This caused libusb_cancel_transfer() to fail, and the read thread would never stop (and hid_close() would never return).
This patch simply sets the shutdown_thread flag from read_close(). This way,
even if hid_close() is called while there is no transfer pending, the loop
in read_thread() will exit, causing read_thread() to end. For this to not
leak resources, read_thread() also now has to call libusb_cancel_transfer().
Note that necessarily, one of the calls to libusb_submit_transfer() will
fail (and return an error code), but that's ok, because one of them is
guaranteed to succeed.

Reported-by: Ludovic Rousseau <ludovic.rousseau@gmail.com>